% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{coeff_table_glmpd}
\alias{coeff_table_glmpd}
\title{coeff_table_glmpd}
\usage{
coeff_table_glmpd(fits)
}
\arguments{
\item{fits}{A List of model fits. output of getProjectionDrivers.R}
}
\value{
A list of coefficient tables
}
\description{
coeff_table_glmpd
}
