% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{.fitGLM}
\alias{.fitGLM}
\title{.fitGLM
This function parallelizes glm models:
Parallelizes across genes in a counts matrix
Parallelizes across patterns in a matrix of projected patterns}
\usage{
.fitGLM(
  countsMatrix,
  annotDF,
  model_formula_str,
  projected_patterns,
  cores,
  geneDF = NULL
)
}
\arguments{
\item{countsMatrix}{matrix of gene expression counts. genes x cells / samples.}

\item{annotDF}{A dataframe with annotations for each cell or sample in the gene expression dataset}

\item{model_formula_str}{A string specifying the model to fit, in the format "~ var1 + var2"}

\item{projected_patterns}{A matrix of projected pattern weights to use as an explanatory variable in the model fit for each gene. colnames are patterns, rownames are cells. Must have viable column names.}

\item{cores}{Integer defining the number of cores to use for parallelization of model fitting across genes.}

\item{geneDF}{(currently) optional dataframe with annotations for each gene in the dataset}

\item{exp_family}{The regression family to use (default: 'negbinomial')}

\item{clean_model}{Boolean.  Passed through to monocle3::fit_model}

\item{verbose}{Boolean. Verbose output for model fitting.}

\item{result}{Return full or summarized results. Summarized results are stripped of memory-intensive models and return coefficients from extractCoefficients().}
}
\value{
Returns a list of fitted models (output similar to monocle3::fit_model)
}
\description{
.fitGLM
This function parallelizes glm models:
Parallelizes across genes in a counts matrix
Parallelizes across patterns in a matrix of projected patterns
}
