% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{extractCoeffHelper_glm_nb}
\alias{extractCoeffHelper_glm_nb}
\title{extractCoeffHelper_glm_nb}
\usage{
extractCoeffHelper_glm_nb(model, model_summary)
}
\arguments{
\item{model}{A fitted GLM model}

\item{model_summary}{Not sure what this argument is offhand (Alina?)}
}
\value{

}
\description{
extractCoeffHelper_glm_nb
}
