% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{extractCoefficients}
\alias{extractCoefficients}
\title{extractCoefficients
Extract coefficients from model fitting object. Each object of list requires a column exactly named "gene_id".}
\usage{
extractCoefficients(glm_models, genes)
}
\arguments{
\item{glm_models}{Named List (over patterns) of objects returned from "full_result" of fit_models_cds. Organizes each pattern by gene.}

\item{genes}{gene ids that exactly match the count matrix which was provided to fitGLMpd()}
}
\value{
Returns a mutileveled list of fitted model coefficients. Hierarchy of organization: pattern (list), gene (list), coefficients (data.frame)
}
\description{
extractCoefficients
Extract coefficients from model fitting object. Each object of list requires a column exactly named "gene_id".
}
