% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_helper}
\alias{fit_helper}
\title{fit_helper}
\usage{
fit_helper(thisGene, thisPattern, model_formula_string, annotDF)
}
\arguments{
\item{thisGene}{String of gene name}

\item{thisPattern}{String of pattern name}

\item{model_formula_string}{Model Formula String}

\item{annotDF}{Not sure what this argument is (alina?)}
}
\value{

}
\description{
fit_helper
}
