% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_model_cds}
\alias{fit_model_cds}
\title{Fit Models (cds)
This function wraps around monocle3::fit_model to parallelize glm models for a given pattern `pattern` across genes in a cds object}
\usage{
fit_model_cds(
  cds,
  pattern,
  cores,
  exp_family = "negbinomial",
  clean_model = F,
  verbose = T
)
}
\arguments{
\item{cds}{A monocle3 cell_data_set object}

\item{pattern}{A vector of projected pattern weights to use as an explanatory variable in the model fit for each gene.}

\item{cores}{Integer defining the number of cores to use for parallelization of model fitting across genes.}

\item{exp_family}{The regression family to use (default: 'negbinomial')}

\item{clean_model}{Boolean.  Passed through to monocle3::fit_model}

\item{verbose}{Boolean. Verbose output?}
}
\value{
Returns a list of fitted models (output similar to monocle3::fit_model)
}
\description{
Fit Models (cds)
This function wraps around monocle3::fit_model to parallelize glm models for a given pattern `pattern` across genes in a cds object
}
