% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fitGLMpd}
\alias{fitGLMpd}
\title{fitGLMpd
Uses a GLM to identify the set of genes used in a pattern for a given cell type or condition}
\usage{
fitGLMpd(object, ..., verbose = TRUE)
}
\arguments{
\item{verbose}{Boolean. Verbose output for model fitting.}

\item{countsMatrix}{matrix of gene expression counts. genes x cells / samples.}

\item{annotDF}{A dataframe with annotations for each cell or sample in the gene expression dataset}

\item{model_formula_str}{A string specifying the model to fit, in the format "~ var1 + var2"}

\item{projected_patterns}{A matrix of projected pattern weights to use as an explanatory variable in the model fit for each gene. colnames are patterns, rownames are cells. Must have viable column names.}

\item{cores}{Integer defining the number of cores to use for parallelization of model fitting across genes.}

\item{exp_family}{The regression family to use (default: 'negbinomial')}

\item{clean_model}{Boolean.  Passed through to monocle3::fit_model}

\item{result}{Return full or summarized results. Summarized results are stripped of memory-intensive models and return coefficients from extractCoefficients().}
}
\value{
Returns a list of fitted models (output similar to monocle3::fit_model)
}
\description{
fitGLMpd
Uses a GLM to identify the set of genes used in a pattern for a given cell type or condition
}
