% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{getProjectionDrivers}
\alias{getProjectionDrivers}
\title{Title}
\usage{
getProjectionDrivers(countsMatrix, annotDF, projectedPs, model_formula_str)
}
\arguments{
\item{countsMatrix}{genes x samples matrix with gene counts}

\item{annotDF}{samples x annotations matrix. includes condition(s) of interest}

\item{projectedPs}{patterns x sample matrix with projected p weights (learned with projectR)}

\item{model_formula_str}{}
}
\value{
A list of model fits for each pattern. Each item in that list contains a fit for each gene
}
\description{
Title
}
