% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{organizeEstimates}
\alias{organizeEstimates}
\title{organizeEstimates
Generate matrices of coefficients of interest for later plotting with heatmaps.}
\usage{
organizeEstimates(
  coefficients_list,
  terms_exact,
  terms_match,
  feature = "estimate",
  gene_name = "gene_id",
  transpose = F
)
}
\arguments{
\item{coefficients_list}{List (over patterns) of objects returned from orderCoefficientByGene()}

\item{terms_exact}{String. Names of terms to plot which exactly match the terms of the model.}

\item{terms_match}{String. specify beginning string to match terms to plot. Necessary for interaction terms.}

\item{feature}{which feature to pull from the model. Default ["estimate"] is beta. Options include "q_value"...}

\item{gene_name}{name of column which contains gene names. Default ["gene_id"]}

\item{tranpose}{boolean. Should matrices be tranposed. Default [F] returns genes as rows.}
}
\value{
Returns a list of matrices.
}
\description{
organizeEstimates
Generate matrices of coefficients of interest for later plotting with heatmaps.
}
