% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCellPatterns}
\alias{plotCellPatterns}
\title{plotCellPatterns
This function provides a wrapper around monocle3's plotting function to allow for clipping of color scales, to limit the effect of outliers on visualization.
Implemented to plot continuous variables in pData that share common prefixes and unique appended ids... eg "pattern1" "pattern2" "patternZ"}
\usage{
plotCellPatterns(
  pattern_suffix,
  cds,
  pattern_prefix = NULL,
  red_method = "UMAP",
  do.clip = c(0, 1)
)
}
\arguments{
\item{pattern_suffix}{A vector of names (or unique parts of names - used with pattern_prefix) of continuous values in pData to color cells by}

\item{cds}{A monocle3 cell_data_set object}

\item{pattern_prefix}{Used with pattern_suffix, if all variables to be plotted have a common prefix. only implemented like this to allow for lapply}

\item{red_method}{Passed through monocle::plot_cells}

\item{do.clip}{A vector of two numbers on the scale [0,1] indicating minimum and maximum percentiles at which to clip the color scale.}
}
\value{
Returns a plot (or list of plots) with cells colored by pattern weight
}
\description{
plotCellPatterns
This function provides a wrapper around monocle3's plotting function to allow for clipping of color scales, to limit the effect of outliers on visualization.
Implemented to plot continuous variables in pData that share common prefixes and unique appended ids... eg "pattern1" "pattern2" "patternZ"
}
