% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPatternUsageByCondition}
\alias{plotPatternUsageByCondition}
\title{#Function to plot a continuous variable (eg pattern usage) over a categorical variable such as time.
Provide feature (continuous) and bin_by (categorical, discrete) as strings that matches a column in pData.}
\usage{
plotPatternUsageByCondition(feature, cds, conditions, bin_by = NULL)
}
\arguments{
\item{feature}{A names or vector of names of continuous values in pData to plot the values}

\item{cds}{A monocle3 cell_data_set object}

\item{bin_by}{A categorical value in pData to group cells by}
}
\value{
Returns a boxplot (or list of plots) of cells projected pattern weights, for groups of interest
}
\description{
#Function to plot a continuous variable (eg pattern usage) over a categorical variable such as time.
Provide feature (continuous) and bin_by (categorical, discrete) as strings that matches a column in pData.
}
