% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{resid}
\alias{resid}
\title{Residual Plot}
\usage{
resid(model, gene)
}
\arguments{
\item{model}{A model fit using fit_models}

\item{gene}{A string of gene name (must be in <where is this matching?>)}
}
\value{

}
\description{
Residual Plot
}
\examples{
<add example code here for how to use this function>
}
