% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{extractCoefficients}
\alias{extractCoefficients}
\title{Extract coefficients from model fitting object}
\usage{
extractCoefficients(glm_models, genes)
}
\arguments{
\item{glm_models}{List (over patterns) of objects returned from "full_result" of fit_models_cds. Organizes each pattern by gene.}

\item{genes}{ensembl gene ids that exactly match those which were provided to fit_models_cds}
}
\value{
Returns a list of fitted models (output similar to monocle3::fit_model)
}
\description{
Extract coefficients from model fitting object
}
