% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{orderCoefficientsByGene}
\alias{orderCoefficientsByGene}
\title{Order model coefficient estimates by genes, such as for visualization by heatmap}
\usage{
orderCoefficientsByGene(
  pattern_coefficient_list,
  coefficients_to_keep = NULL,
  filter_significance,
  string
)
}
\arguments{
\item{pattern_coefficient_list}{List (over patterns) of objects returned from "summarized_result" of fit_models_cds(), or extractCoefficients().}

\item{coefficients_to_keep}{optional. limit returned dataframe to coefficients of interest.}

\item{filter_significance}{Numeric value at which to q-value significance. If not provided, all genes returned.}

\item{string}{optional. all parameters that contain this string will be used for signficance filtering.}
}
\value{
Returns a list of fitted models (output similar to monocle3::fit_model)
}
\description{
Order model coefficient estimates by genes, such as for visualization by heatmap
}
